Added Staff features in admin panel
Added some missing features to the Admin Panel when editing a member of staff.
These features were available to the agent when they logged in under "Profile" but were missing on the admin panel.

This can be very handy to see what agents have set for their preferences as well as being able to change them.

Also added the ability to edit the agents signature as HTML as this was missing in the admin panel and the staff panel.

Added the ability to edit the department signature as HTML as this was also missing.

Added the ability to edit the Authentication and Registration Templates & Pages within the "Settings > Agents > Templates" screen of the admin panel as this was also missing.