fix(email): fix case sensitivity in "Action" header retrieval
This fixes a casing issue in the handling of the Action header while handling bounce notices from mail servers.

Prior to this change, no emails could be flagged as a bounce email because the Upper-Case version of Action was used. This has been lowercases to "action".

An accompanying unit test, tests that a bounce email is properly passed and it's attributes are setup. I have redacted a sample DSN email as much as I thought was necessary but am happy to re-review this fixture.

As all other header access appears to be lowercase, this just appears to be a simple typo.

This fixes: https://forum.osticket.com/d/104433-undelivered-mail-returned-to-sender-processing too